+++ src/IB/KrylovMobilitySolver.cpp
@@ -994,8 +994,8 @@ KrylovMobilitySolver::monitorKSP(KSP ksp, int it, PetscReal rnorm, void* /*mctx*
     }
 
     std::streamsize old_precision = tbox::plog.precision(16);
-    tbox::plog << std::scientific << it << " KMInv_KSP " << print_normtype << " resid norm " << static_cast<double>(rnorm)
-               << " true resid norm " << static_cast<double>(truenorm) << " ||r(i)||/||b|| " << static_cast<double>(truenorm / bnorm)
+    tbox::plog << std::scientific << it << " KMInv_KSP " << print_normtype << " resid norm " << rnorm
+               << " true resid norm " << truenorm << " ||r(i)||/||b|| " << truenorm / bnorm
                << std::endl;
 
     tbox::plog.precision(old_precision);
